==> Code/1.0 Data Pull.Rmd <==
---
title: "ACS Data"
author: "Madeline Chang"
date: "2025-07-01"
output: pdf_document
editor_options: 
  chunk_output_type: console
---

Resources:
https://resources.github.com/github-and-rstudio/

https://api.census.gov/data/key_signup.html



## Calling in libraries

```{r setup, include=FALSE}
library(dplyr)
library(tidyverse)
library(ggplot2)
library(janitor)
library(lubridate)
library(googlesheets4)
library(rlang)
library(gridExtra)
library(corrplot)
library(forcats)
library(stats)


library(tidycensus)
```

----------------------


```{r}
# PUMS data example (takes a while to run, and we would pull in smaller batches of columns)

pums_var<- pums_variables %>% 
  filter(year == 2023, survey == "acs5") %>%
  distinct(var_code, var_label, data_type, level) %>% 
  filter(level == "person")

pums_variables %>% 
  filter(survey == "acs5") %>%
  distinct(var_code, var_label, data_type, level) %>%
  filter(var_code == 'SERIALNO')
```

4-14: citizenship, year of naturalization write in, class of worker, difficulties, veteran service connected disability rating (% and check box)
20-26: insurance
59: educational attainment
61: sex
62-65: wages/salary income, hours worked per week in past year
74: disabled status
77: employment status
96: total person's income
98: income to poverty ratio re-code
108-111, 113: race

```{r}
vars<- c(pums_var$var_code[4:14], pums_var$var_code[20:26], pums_var$var_code[59], pums_var$var_code[61:65], pums_var$var_code[74], pums_var$var_code[77], pums_var$var_code[96], pums_var$var_code[98],pums_var$var_code[108:111], pums_var$var_code[113])

pums_data<- get_pums(variables = c('AGEP', 'PUMA'), variables_filter = list(AGEP = 20:65), year = 2023, state = "IL", survey = "acs5", recode = TRUE)

pums_data<- pums_data %>%
  filter(substr(SERIALNO, 1, 6) == '2023HU') %>%
  select(SERIALNO, SPORDER, AGEP, WGTP, PWGTP, PUMA)

for (var in vars){
  var_data<- get_pums(variables = c(var, 'PUMA'), variables_filter = list(AGEP = 20:65), year = 2023, state = "IL", survey = "acs5", recode = TRUE) %>%
    filter(substr(SERIALNO, 1, 6) == '2023HU') %>%
    select(SERIALNO, SPORDER, starts_with(var), PUMA)
  
  pums_data<- pums_data %>%
    left_join(var_data, by = join_by(SERIALNO, SPORDER, PUMA))
  
  Sys.sleep(10)
  
  print(paste("Data pull for", var,"is complete"))
}
```

```{r}
write_csv(pums_data, '/Users/mtc/ADS/ADS 599/599 Project/pums_data.csv')
```
