==> Code/2.0 EDA.Rmd <==
---
title: "ADS 599 Capstone: EDA"
author: "Madeline Chang"
date: "2025-11-01"
output: pdf_document
editor_options: 
  chunk_output_type: console
---

## Calling in libraries

```{r setup, include=FALSE}
library(dplyr)
library(tidyverse)
library(ggplot2)
library(janitor)
library(lubridate)
library(caret)
library(rlang)
library(gridExtra)
library(corrplot)
library(forcats)
library(stats)
library(mosaic)
library(knitr)
library(weights)
library(Hmisc)
library(corrplot)
library(here)
library(httr)

library(tidycensus)
```

## Reading in the data 

```{r setup, include=FALSE}
git_pums_data <- GET('https://raw.githubusercontent.com/mchang674/ADS599_Capstone/refs/heads/main/Data%20Files/pums_data.csv') 

git_puma_data<- GET('https://raw.githubusercontent.com/mchang674/ADS599_Capstone/refs/heads/main/Data%20Files/2020_Census_Tract_to_2020_PUMA.txt')


pums_data<- readr::read_csv(git_pums_data)
puma<- readr::read_csv(git_puma_data)
```

## Adjusting indicator variables so that No = 0 and Yes = 1

```{r}
bin_fix<- c('DDRS', 'DEAR', 'DEYE', 'DOUT', 'DPHY', 'DREM', 'HINS1', 'HINS2', 'HINS3', 'HINS4', 'HINS5', 'HINS6', 'HINS7', 'DIS')

for(label in bin_fix){
  pums_data[[label]] <- ifelse(pums_data[[label]] == 2, 0, 1)
}
```


## Creating derived variables

```{r}
pums_data_col<- pums_data %>%
  mutate(edu_cat = case_when(SCHL == '01' | SCHL =='02' | SCHL =='03'| SCHL =='04'| SCHL =='05'| SCHL =='06'| SCHL =='07'| SCHL =='08' ~1,
                                SCHL == '09'| SCHL =='10'| SCHL =='11'~2,
                                SCHL == '12'| SCHL =='13'| SCHL =='14'| SCHL =='15'~3,
                                SCHL == '16'| SCHL =='17'~4,
                                SCHL == '18'| SCHL =='19'~5,
                                SCHL == '20'| SCHL =='21'~6,
                                SCHL == '22'| SCHL =='23'| SCHL =='24'~7),
         COW_desc = case_when(COW == 1 ~ "for_profit_or_ind",
                              COW == 2 ~ "not_profit",
                              COW == 3 ~ "local_gov",
                              COW == 4 ~ "state_gov",
                              COW == 5 ~ "fed_gov",
                              COW == 6 ~ "self_not_incorp",
                              COW == 7 ~ "self_incorp",
                              COW == 8 ~ "family_bus_wo_pay",
                              COW == 9 ~ "unemp",
                              COW == 'b' ~ "not_labor_force"),
         below_poverty = as.integer(POVPIP < 100),
         deep_poverty = as.integer(POVPIP < 50)) %>%
  filter(!POVPIP == -1)
```


# Adding geography information
```{r}
tract_per_puma<- puma %>%
  filter(STATEFP == '17') %>% 
  group_by(PUMA5CE) %>%
  summarise(tract_in = length(unique(TRACTCE))) 


pums_data_col<- pums_data_col %>%
  inner_join(tract_per_puma, by = join_by(PUMA == PUMA5CE))

```


# Total Income and Earnings

## Summary statistics (weighted) of total income by sex, disability, and education attainment

```{r}
pums_data_col %>%
  group_by(SEX_label, DIS_label, edu_cat) %>%
  summarise(min = wtd.quantile(PINCP, weight = PWGTP, probs=c(0)),
            q1 = wtd.quantile(PINCP, weight = PWGTP, probs=c(0.25)),
            median = wtd.median(PINCP, weight = PWGTP),
            q3 = wtd.quantile(PINCP, weight = PWGTP, probs=c(0.75)),
            max = wtd.quantile(PINCP, weight = PWGTP, probs=c(1)),               
            mean = weighted.mean(PINCP, w=PWGTP),
            count = sum(PWGTP)
) %>%
  kable()
```

## Weighted boxplots of total income and earnings in the last 12 months by educational attainment and disability

```{r}
ggplot(data = pums_data_col, weight = PWGTP) +
  geom_boxplot(aes(x = edu_cat, y = PINCP, fill = DIS_label)) +
  labs(x = 'Educational Attainment', y = 'Total Income', fill = "Disability Status", title = "Total Income by Educational Attainment")
```


## Weighted boxplots of total income and earnings in the last 12 months by educational attainment and disability

```{r}
ggplot(data = pums_data_col, weight = PWGTP) +
  geom_boxplot(aes(x = edu_cat, y = POVPIP, fill = DIS_label)) +
  labs(x = 'Educational Attainment', y = '% of Poverty Line', fill = "Disability Status", title = "Total Income by Percent of Poverty Line")

ggplot(data = pums_data_col, weight = PWGTP) +
  geom_boxplot(aes(x = edu_cat, y = WAGP, fill = DIS_label)) +
  labs(x = 'Educational Attainment', y = 'Income from Last 12 Months', fill = "Disability Status", title = "Income from Last 12 Months by Educational Attainment")
```


# Percent of Poverty Line

```{r}
pums_data_col %>%
  group_by(SEX_label, DIS_label, edu_cat) %>%
  summarise(min = wtd.quantile(POVPIP, weight = PWGTP, probs=c(0)),
            q1 = wtd.quantile(POVPIP, weight = PWGTP, probs=c(0.25)),
            median = wtd.median(POVPIP, weight = PWGTP),
            q3 = wtd.quantile(POVPIP, weight = PWGTP, probs=c(0.75)),
            max = wtd.quantile(POVPIP, weight = PWGTP, probs=c(1)),               
            mean = weighted.mean(POVPIP, w=PWGTP),
            count = sum(PWGTP)
) %>%
  kable()
```


## Average Poverty Rate (weighted) by Education and Disability poverty table education & disability
```{r}
poverty_by_group <- pums_data_col %>%
  group_by(DIS_label, edu_cat) %>% 
  summarise(
    poverty_rate = weighted.mean(below_poverty, w = PWGTP, na.rm = TRUE),
    deep_poverty_rate = weighted.mean(deep_poverty,  w = PWGTP, na.rm = TRUE),
    n_weighted = sum(PWGTP, na.rm = TRUE),
    .groups = "drop"
  )

knitr::kable(poverty_by_group, digits = 3,
             caption = "Weighted poverty rates by disability status and education")
```

## Heatmap of Poverty Rate by Education and Disability
```{r fig.width=7, fig.height=4}
ggplot(poverty_by_group, aes(x = edu_cat, y = DIS_label, fill = poverty_rate)) +
  geom_tile(color = "white") +
  scale_fill_viridis_c(labels = scales::percent, name = "Poverty rate") +
  labs(
    title = "Poverty by Education and Disability Status - Weighted",
    x = "Education obtained", y = "Disability status"
  ) +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))
```

## Heatmap of Deep Poverty Rate by Education and Disability
```{r}
ggplot(poverty_by_group, aes(x = edu_cat, y = DIS_label, fill = deep_poverty_rate)) +
  geom_tile(color = "white") +
  scale_fill_viridis_c(labels = scales::percent, name = "Deep poverty rate") +
  labs(
    title = "Deep Poverty by Education and Disability Status (Weighted)",
    x = "Education obtained", y = "Disability status"
  ) +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))
```

## Visualizing Poverty rates with Increasing Education Level and Disability
```{r}
ggplot(poverty_by_group, aes(x = edu_cat, y = poverty_rate, color = DIS_label, group = DIS_label)) +
  geom_line(linewidth = 1.2) +
  geom_point(size = 2) +
  scale_y_continuous(labels = scales::percent_format()) +
  labs(
    title = "Weighted Poverty Rate by Education and Disability Status",
    x = "Education Level", y = "Poverty Rate", color = "Disability Status"
  ) +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))
```

```{r}
poverty_by_group %>%
  group_by(DIS_label) %>%
  summarise(mean_poverty = weighted.mean(poverty_rate, w = n_weighted))
```
DIS_label            mean_poverty
With a disability           0.229
Without a disability        0.903


# Disability Difficulty

## Correlation between Disability Difficulties

```{r}
pums_data_col %>%
  mutate(num_diff = DDRS + DEAR + DEYE + DOUT + DPHY + DREM) %>%
  group_by(num_diff) %>%
  summarise(count = sum(PWGTP)) %>%
  kable()

pums_data_col %>%
  select(DDRS, DEAR, DEYE, DOUT, DPHY, DREM) %>%
  cor() %>%
  corrplot(diag = FALSE, type = 'upper')
```


## Barplots of Disabilty Difficulty by Educational Attainment

```{r}
disabled<- pums_data_col %>%
  pivot_longer(cols = c(DDRS_label, DEAR_label, DEYE_label, DOUT_label, DPHY_label, DREM_label), names_to = "difficulty", values_to = "value") %>%
  filter(value == "Yes") %>%
  mutate(difficulty = case_when(difficulty == "DDRS_label"~ "self care",
                                difficulty == "DEAR_label" ~ "hearing",
                                difficulty == "DEYE_label" ~ "vision",
                                difficulty == "DOUT_label" ~ "independent living",
                                difficulty == "DPHY_label"~ "ambulatory",
                                difficulty == "DREM_label"~ "cognitive"))

disabled %>%
  ggplot(weight = PWGTP) +
  geom_bar(aes(x = difficulty, fill = edu_cat), position = "dodge") +
  labs(x = "Difficulty", y = "Count", fill = "Educational Attainment", title = "Count of Disability Difficulty by Educational Attainment")

```

```{r}
disabled %>%
  ggplot(weight = PWGTP) +
  geom_boxplot(aes(x = difficulty, y = POVPIP))
```


## Barplots of Disabilty Difficulty by Type of Work

```{r}
disabled %>%
  filter(difficulty == "self care") %>%
  ggplot(weight = PWGTP) +
  geom_bar(aes(x = COW_desc, fill =SEX_label), position = "dodge") +
  labs(x = "Type of Work", y = "Count", fill = "Sex", title = "Count of Type of Work with Self Care Difficulty") +
  coord_flip()

disabled %>%
  filter(difficulty == "hearing") %>%
  ggplot(weight = PWGTP) +
  geom_bar(aes(x = COW_desc, fill =SEX_label), position = "dodge") +
  labs(x = "Type of Work", y = "Count", fill = "Sex", title = "Count of Type of Work with Hearing Difficulty") +
  coord_flip()

disabled %>%
  filter(difficulty == "vision") %>%
  ggplot(weight = PWGTP) +
  geom_bar(aes(x = COW_desc, fill =SEX_label), position = "dodge") +
  labs(x = "Type of Work", y = "Count", fill = "Sex", title = "Count of Type of Work with Vision Difficulty") +
  coord_flip()

disabled %>%
  filter(difficulty == "ambulatory") %>%
  ggplot(weight = PWGTP) +
  geom_bar(aes(x = COW_desc, fill =SEX_label), position = "dodge") +
  labs(x = "Type of Work", y = "Count", fill = "Sex", title = "Count of Type of Work with Ambulatory Difficulty") +
  coord_flip()

disabled %>%
  filter(difficulty == "independent living") %>%
  ggplot(weight = PWGTP) +
  geom_bar(aes(x = COW_desc, fill =SEX_label), position = "dodge") +
  labs(x = "Type of Work", y = "Count", fill = "Sex", title = "Count of Type of Work with Independent Living Difficulty") +
  coord_flip()

disabled %>%
  filter(difficulty == "cognitive") %>%
  ggplot(weight = PWGTP) +
  geom_bar(aes(x = COW_desc, fill =SEX_label), position = "dodge") +
  labs(x = "Type of Work", y = "Count", fill = "Sex", title = "Count of Type of Work with Cognitive Difficulty") +
  coord_flip()

```

# Educational Attainment

## Barplots by Educational Attainment

```{r}
disabled %>%
  mutate(SCHL = as.numeric(SCHL)) %>%
  filter(SCHL >= 1 & SCHL <= 15) %>%
  ggplot(weight = PWGTP) +
  geom_bar(aes(x = edu_cat, fill = COW_desc), position = "dodge") +
  labs(x = "Educational Attainment", y = "Count", title = "Educational Attainment by Type of Work for Disabled Individuals")

disabled %>%
  mutate(SCHL = as.numeric(SCHL)) %>%
  filter(SCHL >= 16 & SCHL <= 19) %>%
  ggplot(weight = PWGTP) +
  geom_bar(aes(x = edu_cat, fill = COW_desc), position = "dodge") +
  labs(x = "Educational Attainment", y = "Count", title = "Educational Attainment by Type of Work for Disabled Individuals")

disabled %>%
  mutate(SCHL = as.numeric(SCHL)) %>%
  filter(SCHL > 19) %>%
  ggplot(weight = PWGTP) +
  geom_bar(aes(x = edu_cat, fill = COW_desc), position = "dodge") +
  labs(x = "Educational Attainment", y = "Count", title = "Educational Attainment by Type of Work for Disabled Individuals")
```


## Correlation of All Numeric Columns

```{r}
pums_data_col %>%
  select(AGEP, DDRS, DEAR, DEYE, DOUT, DPHY, DREM, SSIP, SSP, WAGP, PINCP, tract_in, edu_cat, POVPIP) %>%
  cor() %>%
  corrplot(diag = FALSE, type = 'upper')
```



# Data Pre-Processing for Modeling

```{r}
model_data<- pums_data_col %>%
  select(PWGTP, POVPIP, AGEP, edu_cat, COW_desc, DIS, RACAIAN, RACASN, RACBLK, RACNH, 
         RACPI, WKHP, SEX, HINS1, HINS2, HINS3, HINS4, HINS5, HINS6, HINS7, DDRS, DEAR, 
         DEYE, DOUT, DPHY, DRAT, DRATX, DREM, tract_in) %>%
  rename(ins_prev_work = HINS1,
         ins_direct = HINS2,
         ins_medicare = HINS3,
         ins_medicaid = HINS4,
         ins_tricare = HINS5,
         ins_va = HINS6,
         ins_ind_health = HINS7,
         age = AGEP,
         dis = DIS,
         sex = SEX,
         race_indigenous = RACAIAN,
         race_asian = RACASN,
         race_black = RACBLK,
         race_hawaii = RACNH,
         race_pac_isl = RACPI,
         vet = DRATX) %>%
  mutate(male = ifelse(sex == 1, 1, 0),
         pov_ind = case_when(POVPIP < 100 ~ 'pov',
                             POVPIP >= 100 ~ 'not_pov')) %>%
  mutate(pov_ind = as.factor(pov_ind),
         vet = ifelse(vet == 'b', 0, 1)) %>%
  select(-sex)
  
```

```{r}
write_csv(model_data, '/Users/mtc/ADS/ADS 599/599 Project/model_data.csv')
```
